Refactor settings tests by engine
